Time: 48 ms (64.07%), Space: 42 MB (19.11%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -35,7 +35,7 @@ class Solution {
             }
         }
 
-        return (int)dp[n][m];
+        return dp[n][m];
     }
 
 public:
